pyproject.toml
[project]
name = "whisper-transcription-tool"
version = "0.1.0"
description = "A comprehensive Python application for audio recording, transcription, and processing using OpenAI's Whisper API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "pydub>=0.25.1",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0",
    "click>=8.0.0"
]
